requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
drf-yasg
psycopg2-binary
gunicorn
python-dotenv